% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialMinRand.R
\name{imbalance}
\alias{imbalance}
\title{Treatment Imbalance}
\usage{
imbalance(treatments, treatment.ratios = setNames(rep(1, length(treatments)),
  levels(treatments)))
}
\arguments{
\item{treatments}{a factor containing a sequence of treatments.}

\item{treatment.ratios}{a named vector containing the desired treatment ratios.
By default, each treatment is given in the same frequency.}
}
\description{
Provides a measure of treatment imbalance with respect to a desired ratio of
treatments.
}
\details{
A treatment vector has imbalance 0 if the treatments occur in the desired ratio.
Otherwise, the imbalance is positive.
}
\examples{
# A balanced treatment vector
imbalance(factor(c("A","B","B","A"),levels=c("A","B")))
# Still balanced if ratio is given as 50:50 instead of the default 1:1
imbalance(factor(c("A","B","B","A"),levels=c("A","B")),treatment.ratios=c(A=50,B=50))

}

